Rename config_files.py to file.py
